Troisieme version non complete

+++ classe template/test_polynome.cpp
@@ -3,15 +3,15 @@
 
 int main()
 {
-    Polynome<float> p(3);
+    Polynome<int> p(3);
     p[0] = 2;
     p[1] = 3;
     p[2] = 1;
     p[3] = -1;
 
-    Polynome<float>q({1,2,5,4});
+    Polynome<int>q({1,2,5,4});
 
-    Polynome<float> r(p+q);
+    Polynome<int> r(p+q);
     std::cout << "r = " << r << std::endl;
     for ( int i = 0; i < 10; ++i )
     {
